+++ ESFE.BusinessLogic/UseCases/Products/Queries/GetProduct/GetProductQuery.cs
@@ -3,4 +3,4 @@ using MediatR;
 
 namespace ESFE.BusinessLogic.UseCases.Products.Queries.GetProduct;
 
-public record GetProductQuery(long productId) : IRequest<ProductByIdResponse>;
+public record GetProductQuery(int productId) : IRequest<ProductByIdResponse>;
